+++ contrib/Stats.java
@@ -28,8 +28,8 @@ public class Stats
 	final double NICK_LENGTH = 100; // "Significant" lines in WMA calculations.
 	final double CHAN_LENGTH = 1000;
 	final double THRESHOLD = 0.005; // val * THRESHOLD is considered too small to be a part of WMA.
-	final double NICK_ALPHA = Math.exp(Math.log(THRESHOLD) / (double)NICK_LENGTH);
-	final double CHAN_ALPHA = Math.exp(Math.log(THRESHOLD) / (double)CHAN_LENGTH);
+	final double NICK_ALPHA = Math.exp(Math.log(THRESHOLD) / NICK_LENGTH);
+	final double CHAN_ALPHA = Math.exp(Math.log(THRESHOLD) / CHAN_LENGTH);
 
 	public String[] info()
 	{
@@ -167,10 +167,10 @@ public class Stats
 		}
 
 		try {
-			update( "captuation", mes, (double)apiCaptuation( content ) );
+			update( "captuation", mes, apiCaptuation( content ) );
 			int wc = apiWordCount( content );
-			update( "wordcount", mes, (double)wc );
-			update( "characters", mes, (double)apiLength( content ) );
+			update( "wordcount", mes, wc );
+			update( "characters", mes, apiLength( content ) );
 			if (wc > 0)
 				update( "wordlength", mes, apiWordLength( content ) );
 			update( "referred", mes, referred ? 1.0 : 0.0 );
@@ -197,7 +197,7 @@ public class Stats
 				irc.sendContextReply( mes, "Sorry, cannae find datta one." );
 			} else {
 				obj = ret.get(0);
-				irc.sendContextReply( mes, "They be 'avin a score of " + (double)Math.round(obj.value * 100) / 100.0 );
+				irc.sendContextReply( mes, "They be 'avin a score of " + Math.round(obj.value * 100) / 100.0 );
 			}
 		} else if (params.length == 2) {
 			String nick = mods.nick.getBestPrimaryNick( params[1] );
@@ -207,7 +207,7 @@ public class Stats
 			} else {
 				StringBuilder results = new StringBuilder( "Stats:" );
 				for (EntityStat obj: ret) {
-					results.append( " " + obj.statName + " = " + (double)Math.round(obj.value * 100) / 100.0 + ";" );
+					results.append( " " + obj.statName + " = " + Math.round(obj.value * 100) / 100.0 + ";" );
 				}
 				irc.sendContextReply( mes, results.toString() );
 			}
